feat: further simplifying layouts/pages
