+++ frontend/webapp/app/(overview)/actions/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
 import React from 'react';
-import { OverviewFocusedActions } from '@/components';
+import { useActionCRUD } from '@/hooks';
+import { ActionTable } from '@odigos/ui-containers';
 
 export default function Page() {
-  return (
-    <>
-      <OverviewFocusedActions />
-    </>
-  );
+  const { actions } = useActionCRUD();
+
+  return <ActionTable actions={actions} tableMaxHeight='calc(100vh - 220px)' />;
 }
