+++ frontend/webapp/app/(overview)/destinations/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
 import React from 'react';
-import { OverviewFocusedDestinations } from '@/components';
+import { useDestinationCRUD } from '@/hooks';
+import { DestinationTable } from '@odigos/ui-containers';
 
 export default function Page() {
-  return (
-    <>
-      <OverviewFocusedDestinations />
-    </>
-  );
+  const { destinations } = useDestinationCRUD();
+
+  return <DestinationTable destinations={destinations} tableMaxHeight='calc(100vh - 220px)' />;
 }
