+++ frontend/webapp/app/(overview)/overview/page.tsx
@@ -1,12 +1,44 @@
 'use client';
 
 import React from 'react';
-import { OverviewMain } from '@/components';
+import type { Source } from '@odigos/ui-utils';
+import { DataFlow, MultiSourceControl, type SourceSelectionFormData } from '@odigos/ui-containers';
+import { useActionCRUD, useDestinationCRUD, useInstrumentationRuleCRUD, useMetrics, useSourceCRUD } from '@/hooks';
 
 export default function Page() {
+  const { metrics } = useMetrics();
+  const { actions, actionsLoading } = useActionCRUD();
+  const { destinations, destinationsLoading } = useDestinationCRUD();
+  const { sources, sourcesLoading, persistSources } = useSourceCRUD();
+  const { instrumentationRules, instrumentationRulesLoading } = useInstrumentationRuleCRUD();
+
   return (
     <>
-      <OverviewMain />
+      <DataFlow
+        heightToRemove='176px'
+        sources={sources}
+        sourcesLoading={sourcesLoading}
+        destinations={destinations}
+        destinationsLoading={destinationsLoading}
+        actions={actions}
+        actionsLoading={actionsLoading}
+        instrumentationRules={instrumentationRules}
+        instrumentationRulesLoading={instrumentationRulesLoading}
+        metrics={metrics}
+      />
+
+      <MultiSourceControl
+        totalSourceCount={sources.length}
+        uninstrumentSources={(payload) => {
+          const inp: SourceSelectionFormData = {};
+
+          Object.entries(payload).forEach(([namespace, sources]: [string, Source[]]) => {
+            inp[namespace] = sources.map(({ name, kind }) => ({ name, kind, selected: false }));
+          });
+
+          persistSources(inp, {});
+        }}
+      />
     </>
   );
 }
