+++ frontend/webapp/components/lib-imports/index.ts
@@ -1,9 +1,4 @@
 export * from './error-boundary';
-export * from './overview-focused-actions';
-export * from './overview-focused-destinations';
-export * from './overview-focused-rules';
-export * from './overview-focused-sources';
 export * from './overview-header';
-export * from './overview-main';
 export * from './overview-modals-and-drawers';
 export * from './setup-header';
